% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.parameterservr.R
\docType{package}
\name{parameterservr}
\alias{parameterservr}
\alias{parameterservr-package}
\title{The parameter server for large-scale sparse logistic regression.}
\description{
The parameter server for large-scale sparse logistic regression.
}

